# add some basic perf check script