--- a/contrib/perf/nslookup.py
+++ b/contrib/perf/nslookup.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+
+import socket
+import sys
+
+for i in range(0, 10_000):
+    socket.getaddrinfo(sys.argv[1], 0)