Using podman spawn some containers and then check the performance from aardvark-dns.

Comparing results from now after the rework with v1.11 I can see significant gains. While the total time is about the same the new version only uses around 2/3 of the cycles which means cpu utilization is much better now.