@Luap99 2024-07-23
new: 
```
 Performance counter stats for process id '1449032':

          3,897.06 msec task-clock:u                     #    0.973 CPUs utilized             
                 0      context-switches:u               #    0.000 /sec                      
                 0      cpu-migrations:u                 #    0.000 /sec                      
                40      page-faults:u                    #   10.264 /sec                      
     6,829,395,289      cycles:u                         #    1.752 GHz                       
     6,851,457,495      instructions:u                   #    1.00  insn per cycle            
     1,395,143,347      branches:u                       #  357.999 M/sec                     
        12,815,753      branch-misses:u                  #    0.92% of all branches           

       4.003890371 seconds time elapsed
```

old (v1.11.0):
```

 Performance counter stats for process id '1448088':

          4,730.64 msec task-clock:u                     #    1.182 CPUs utilized             
                 0      context-switches:u               #    0.000 /sec                      
                 0      cpu-migrations:u                 #    0.000 /sec                      
                32      page-faults:u                    #    6.764 /sec                      
     9,741,298,492      cycles:u                         #    2.059 GHz                       
    14,255,917,727      instructions:u                   #    1.46  insn per cycle            
     2,969,731,831      branches:u                       #  627.765 M/sec                     
        24,753,091      branch-misses:u                  #    0.83% of all branches           

       4.003881880 seconds time elapsed
```


I run it several times and the results were pretty consistent.