@mheon 2024-07-23
Seems like network latency probably dominates the wall time...

Suggestion: Time the first one separately, as it shouldn't have a full DNS cache.