@Luap99 2024-07-23
> Time the first one separately, as it shouldn't have a full DNS cache.

What cache? I don't see how I can time the first one differently.