@mheon 2024-07-23
The coredns server should be caching, no? We'll store the result locally and respond based on it within the limits of the TTL.