@Luap99 2024-07-23
> The coredns server should be caching, no? We'll store the result locally and respond based on it within the limits of the TTL.

There is no cache at all AFAICT and we only measure internal names here so it goes though the normal map lookup each time.