@Luap99 2024-07-23
> benchmarking aardvark-dns as isolated DNS server 

I mean this sort of does that. I don't know how you would test isolated? You need a client sending requests, whenever this is a podman container or not seems irrelevant for the most part. Of course the total time will be different as we do not have to overhead of podman starting/stopping containers but honestly I do not care about that at all.

So far this seems more than good enough for a quick check which is all I care about. I just needed to know whenever my changes make it better or worse. This should not be considered an actual benchmark.