@flouthoc 2024-07-23
> So far this seems more than good enough for a quick check which is all I care about. I just needed to know whenever my changes make it better or worse. This should not be considered an actual benchmark.

Yes I agree this looks good for the quick check.

> I don't know how you would test isolated

Initially when we were developing aardvark I remember checking it `e2e` behavior without `podman` , If I find that hacky arrangement on my system then we can use those bits to measure throughput and other metrics. But lets leave it for another day if we ever decide to have raw bechmarks for `aardvark-dns`.

Overall exisiting perf in this PR LGTM.